Expand GuavaCSC and CaffeineCSC names
